move test config to `code-quality` convention plugin
