+++ buildSrc/src/main/kotlin/buildsrc/conventions/base.gradle.kts
@@ -1,9 +1,5 @@
 package buildsrc.conventions
 
-import buildsrc.utils.Rife2TestListener
-import org.gradle.api.tasks.testing.logging.TestExceptionFormat
-import org.gradle.api.tasks.testing.logging.TestLogEvent
-
 /** common config for all subprojects */
 
 plugins {
@@ -19,13 +15,4 @@ tasks.withType<AbstractArchiveTask>().configureEach {
     // https://docs.gradle.org/current/userguide/working_with_files.html#sec:reproducible_archives
     isPreserveFileTimestamps = false
     isReproducibleFileOrder = true
-}
-
-tasks.withType<Test>().configureEach {
-    val testsBadgeApiKey = providers.gradleProperty("testsBadgeApiKey")
-    addTestListener(Rife2TestListener(testsBadgeApiKey))
-    testLogging {
-        exceptionFormat = TestExceptionFormat.FULL
-        events = setOf(TestLogEvent.PASSED, TestLogEvent.SKIPPED, TestLogEvent.FAILED)
-    }
 }
